+++ src/entities/renderers/PlayerRenderer.cpp
@@ -75,8 +75,11 @@ void PlayerRenderer::draw(SpriteBatch& spriteBatch, const ls::Vec2I& sprites) co
         steppingSpriteVariant = static_cast<int>(distanceTravelled / distanceTravelledPerStep) % numberOfSteppingSprites + 1;
     }
 
-    const ls::Vec2F sprite(sprites.x + steppingSpriteVariant * GameConstants::tileFullSpriteSize, sprites.y + direction * GameConstants::tileFullSpriteSize);
-    const ls::Vec2F size(GameConstants::tileSize, GameConstants::tileSize);
+    const ls::Vec2F sprite(
+        static_cast<float>(sprites.x + steppingSpriteVariant * GameConstants::tileFullSpriteSize), 
+        static_cast<float>(sprites.y + direction * GameConstants::tileFullSpriteSize)
+    );
+    const ls::Vec2F size(static_cast<float>(GameConstants::tileSize), static_cast<float>(GameConstants::tileSize));
     ls::Vec2F pos = m_owner->model().position() + offsetToOrigin;
     pos.x = std::floor(pos.x);
     pos.y = std::floor(pos.y);
